:sparkles: feat: Implement jwt auth on Enterprise login/sign-up
