+++ src/main/kotlin/com/team1/mvp_test/common/Role.kt
@@ -1,0 +1,5 @@
+package com.team1.mvp_test.common
+
+enum class Role {
+    ENTERPRISE, MEMBER, ADMIN
+}
